pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "", email = ""},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "django-environ>=0.10.0",
    "django-cors-headers>=4.2.0",
    "djangorestframework-simplejwt>=5.3.0",
    "setuptools>=68.1.2",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "pylint>=2.13.9",
    "black>=22.3.0",
]
[tool.pdm.scripts]
manage = "python ./manage.py"
server = "python ./manage.py runserver"
makemigrations = "python ./manage.py makemigrations"
migrate = "python ./manage.py migrate"

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"
